+++ deploy/v1.1.0/upgradeBeaconEthereumMainnet.ts
@@ -1,7 +1,7 @@
 import { dim } from 'chalk';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { configureBeaconDeployment } from '../../helpers/upgrade/v1.1.0/configureBeaconDeployment'
-import { handleBeaconContractDeploy } from '../../helpers/upgrade/v1.1.0/handleBeaconContractDeploy'
+import { configureBeaconDeployment } from '../../helpers/upgrade/v1.1.0/upgrade/configureBeaconDeployment'
+import { handleBeaconContractDeploy } from '../../helpers/upgrade/v1.1.0/upgrade/handleBeaconContractDeploy'
 import saveUpgradePopulatedTransactions from '../../helpers/files/saveUpgradePopulatedTransactions';
 
 const upgradeBeaconEthereumMainnet = async (hardhat: HardhatRuntimeEnvironment) => {
