+++ helpers/index.ts
@@ -1,7 +1,2 @@
 export * from './deployContract'
-export * from './configureBeaconChainDeployment'
-export * from './configureReceiverChainDeployment'
-export * from './handleBeaconChainContractDeploy'
-export * from './handlePeripheryContractDeploy'
-export * from './handlePrizePoolCoreDeploy'
-export * from './handleReceiverChainContractDeploy'
+export * from './transferOwnership'
