+++ helpers/upgrade/v1.1.0/new/configureBeaconChainDeployment.ts
