+++ helpers/upgrade/v1.1.0/new/handleBeaconChainContractDeploy.ts
@@ -1,4 +1,4 @@
-import { deployContract } from './deployContract'
+import { deployContract } from '../../../deployContract'
 
 export interface handleBeaconChainContractDeployConfig {
   startingDrawId: string;
