+++ helpers/upgrade/v1.1.0/upgrade/configureBeaconDeployment.ts
