+++ helpers/upgrade/v1.1.0/upgrade/handleBeaconContractDeploy.ts
@@ -1,8 +1,7 @@
-import { deployContract } from '../../deployContract'
+import { deployContract } from '../../../deployContract'
 
-export async function handleReceiverContractDeploy(deploy: Function, deployer: string, ethers: any) {
-  // const prizeTierHistory = await ethers.getContract('PrizeTierHistory')
-  const prizeTierHistory = await deployContract(deploy, 'PrizeTierHistory', deployer, [deployer])
+export async function handleBeaconContractDeploy(deploy: Function, deployer: string, ethers: any) {
+  const prizeTierHistory = await ethers.getContract('PrizeTierHistory')
   const drawBuffer = await ethers.getContract('DrawBuffer')
   const prizeDistributionBuffer = await ethers.getContract('PrizeDistributionBuffer')
   const ticket = await ethers.getContract('Ticket')
@@ -17,15 +16,15 @@ export async function handleReceiverContractDeploy(deploy: Function, deployer: s
     ticket.address,
     1000000 // @NOTE:  1 USDC = 1000000 wei = Minumum ticket cost
   ])
-  const receiverTimelockAndPushRouterResult = await deployContract(deploy, 'ReceiverTimelockAndPushRouter', deployer, [
+
+  const beaconTimelockAndPushRouterResult = await deployContract(deploy, 'BeaconTimelockAndPushRouter', deployer, [
     deployer,
-    drawBuffer.address,
     prizeDistributionFactoryResult.address,
     drawCalculatorTimelock.address
   ])
 
   return {
     prizeDistributionFactory: prizeDistributionFactoryResult,
-    receiverTimelockAndPushRouterResult: receiverTimelockAndPushRouterResult
+    beaconTimelockAndPushRouter: beaconTimelockAndPushRouterResult
   }
 }
