+++ helpers/upgrade/v1.1.0/upgrade/handleReceiverContractDeploy.ts
@@ -1,7 +1,8 @@
-import { deployContract } from '../../deployContract'
+import { deployContract } from '../../../deployContract'
 
-export async function handleBeaconContractDeploy(deploy: Function, deployer: string, ethers: any) {
-  const prizeTierHistory = await ethers.getContract('PrizeTierHistory')
+export async function handleReceiverContractDeploy(deploy: Function, deployer: string, ethers: any) {
+  // const prizeTierHistory = await ethers.getContract('PrizeTierHistory')
+  const prizeTierHistory = await deployContract(deploy, 'PrizeTierHistory', deployer, [deployer])
   const drawBuffer = await ethers.getContract('DrawBuffer')
   const prizeDistributionBuffer = await ethers.getContract('PrizeDistributionBuffer')
   const ticket = await ethers.getContract('Ticket')
@@ -16,15 +17,15 @@ export async function handleBeaconContractDeploy(deploy: Function, deployer: str
     ticket.address,
     1000000 // @NOTE:  1 USDC = 1000000 wei = Minumum ticket cost
   ])
-
-  const beaconTimelockAndPushRouterResult = await deployContract(deploy, 'BeaconTimelockAndPushRouter', deployer, [
+  const receiverTimelockAndPushRouterResult = await deployContract(deploy, 'ReceiverTimelockAndPushRouter', deployer, [
     deployer,
+    drawBuffer.address,
     prizeDistributionFactoryResult.address,
     drawCalculatorTimelock.address
   ])
 
   return {
     prizeDistributionFactory: prizeDistributionFactoryResult,
-    beaconTimelockAndPushRouter: beaconTimelockAndPushRouterResult
+    receiverTimelockAndPushRouterResult: receiverTimelockAndPushRouterResult
   }
 }
