--- src/constants.ts
@@ -1,19 +1,0 @@
-export const DRAW_BUFFER_CARDINALITY = 255;
-export const PRIZE_DISTRIBUTION_BUFFER_CARDINALITY = 180; // six months
-export const BEACON_START_TIME = Math.floor((new Date('2021-10-15T19:00:00.000Z')).getTime() / 1000)
-export const BEACON_PERIOD_SECONDS = 86400; // one day
-export const END_TIMESTAMP_OFFSET = 15 * 60 // 15 minutes
-export const RNG_TIMEOUT_SECONDS = 2 * 3600 // 2 hours
-export const EXPIRY_DURATION = 60 * 86400 // 2 months
-export const TOKEN_DECIMALS = 6;
-
-export default {
-  DRAW_BUFFER_CARDINALITY,
-  PRIZE_DISTRIBUTION_BUFFER_CARDINALITY,
-  BEACON_START_TIME,
-  BEACON_PERIOD_SECONDS,
-  END_TIMESTAMP_OFFSET,
-  RNG_TIMEOUT_SECONDS,
-  EXPIRY_DURATION,
-  TOKEN_DECIMALS
-}
